Updates dashboards of all ac-pkg
