+++ ac/ac-image/src/Dashboard.jsx
@@ -53,7 +53,9 @@ const mergeLog = (data: any, dataFn: Object, log: LogDBT) => {
 const initData = {};
 
 export default {
-  Viewer,
-  mergeLog,
-  initData
+  dashboard: {
+    Viewer,
+    mergeLog,
+    initData
+  }
 };
