+++ ac/ac-quiz/src/Dashboard/index.js
@@ -1,0 +1,15 @@
+// @flow
+
+import { ProgressDashboard, LeaderBoard } from 'frog-utils';
+
+import ScatterDashboard from './ScatterDashboard';
+import CountDashboard from './CountDashboard';
+import JustificationDashboard from './JustificationDashboard';
+
+export default {
+  progress: ProgressDashboard,
+  leaderboard: LeaderBoard,
+  scatter: ScatterDashboard,
+  justification: JustificationDashboard,
+  count: CountDashboard
+};
